requirements.txt
alembic==1.0.6
aniso8601==4.1.0
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Dance==1.3.0
Flask-Login==0.4.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
lazy==1.3
Mako==1.0.7
MarkupSafe==1.1.0
oauthlib==3.0.0
pep8==1.7.1
pycodestyle==2.4.0
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.2.16
SQLAlchemy-Utils==0.33.11
urllib3==1.24.1
URLObject==2.4.3
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
